Removed test code. Moved xmlrpc test code to a separate directory.
